converted SearchView to fragment
