+++ app/src/main/java/io/poundcode/spotifystreamer/base/SpotifyStreamFragment.java
@@ -1,0 +1,34 @@
+package io.poundcode.spotifystreamer.base;
+
+import android.app.Fragment;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import butterknife.ButterKnife;
+
+/**
+ * Created by Atlas on 8/17/2015.
+ */
+public abstract class SpotifyStreamFragment extends Fragment {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getActivity().setTitle(getViewTitle());
+        setHasOptionsMenu(true);
+        setRetainInstance(true);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(getLayoutId(), container, false);
+        ButterKnife.inject(this, view);
+        return view;
+    }
+
+    public abstract int getLayoutId();
+
+    public abstract String getViewTitle();
+}
