+++ app/src/main/java/io/poundcode/spotifystreamer/searching/presenter/SpotifyArtistSearchPresenterImpl.java
@@ -19,14 +19,16 @@ public class SpotifyArtistSearchPresenterImpl implements SpotifySearchPresenter 
 
     public boolean isSearching = false;
     SpotifySearchView<ArtistsPager> mView;
+    Context context;
 
-    public SpotifyArtistSearchPresenterImpl(SpotifySearchView<ArtistsPager> view) {
+    public SpotifyArtistSearchPresenterImpl(SpotifySearchView<ArtistsPager> view, Context context) {
+        this.context = context;
         this.mView = view;
     }
 
     @Override
     public void search(String query) {
-        if (isNetworkConnected((Context) mView)) {
+        if (isNetworkConnected(context)) {
             SpotifyServiceWrapper.getNewService().searchArtists(query, new Callback<ArtistsPager>() {
                 @Override
                 public void success(ArtistsPager artistsPager, Response response) {
